rename word module
