--- src/llm_config_module/providers/__init__.py
@@ -1,11 +1,0 @@
-"""Providers module for LLM Config Module."""
-
-from llm_config_module.providers.base import BaseLLMProvider
-from llm_config_module.providers.azure_openai import AzureOpenAIProvider
-from llm_config_module.providers.aws_bedrock import AWSBedrockProvider
-
-__all__ = [
-    "BaseLLMProvider",
-    "AzureOpenAIProvider",
-    "AWSBedrockProvider",
-]
