+++ src/llm_orchestrator_config/config/__init__.py
@@ -1,7 +1,7 @@
 """Configuration module for LLM Config Module."""
 
-from llm_config_module.config.loader import ConfigurationLoader
-from llm_config_module.config.schema import (
+from llm_orchestrator_config.config.loader import ConfigurationLoader
+from llm_orchestrator_config.config.schema import (
     LLMConfiguration,
     ProviderConfig,
     AzureOpenAIConfig,
