+++ src/llm_orchestrator_config/config/loader.py
@@ -9,16 +9,16 @@ import yaml
 from dotenv import load_dotenv
 from loguru import logger
 
-from llm_config_module.config.schema import (
+from llm_orchestrator_config.config.schema import (
     LLMConfiguration,
     ProviderConfig,
     AzureOpenAIConfig,
     AWSBedrockConfig,
     VaultConfig,
 )
-from llm_config_module.vault.secret_resolver import SecretResolver
-from llm_config_module.types import LLMProvider
-from llm_config_module.exceptions import ConfigurationError, InvalidConfigurationError
+from llm_orchestrator_config.vault.secret_resolver import SecretResolver
+from llm_orchestrator_config.types import LLMProvider
+from llm_orchestrator_config.exceptions import ConfigurationError, InvalidConfigurationError
 
 # Constants
 DEFAULT_CONFIG_FILENAME = "llm_config.yaml"
