+++ src/vector_indexer/vault/__init__.py
@@ -1,12 +1,12 @@
 """Embedding vault module for chunk indexing."""
 
-from chunk_indexing_module.vault.vault_client import EmbeddingVaultClient
-from chunk_indexing_module.vault.secret_resolver import EmbeddingSecretResolver
-from chunk_indexing_module.vault.models import (
+from vector_indexer.vault.vault_client import EmbeddingVaultClient
+from vector_indexer.vault.secret_resolver import EmbeddingSecretResolver
+from vector_indexer.vault.models import (
     AzureEmbeddingSecret,
     get_embedding_secret_model,
 )
-from chunk_indexing_module.vault.exceptions import (
+from vector_indexer.vault.exceptions import (
     EmbeddingVaultError,
     EmbeddingVaultConnectionError,
     EmbeddingVaultSecretError,
