+++ src/vector_indexer/vault/exceptions.py
