+++ src/vector_indexer/vault/models.py
